+++ club/club.tsx
@@ -1,0 +1,58 @@
+import { japi } from '../api/api';
+import { notify, lastEvent } from '../events/events';
+
+type ClubItem = {
+  name: string,
+  count_members: number,
+  search_preference?: boolean,
+};
+
+const sortClubs = (cs: ClubItem[] | undefined) => {
+  const unsortedCs = cs ?? [];
+  const sortedCs = unsortedCs.sort((a, b) => {
+    if (a.name.toLowerCase() > b.name.toLowerCase()) return +1;
+    if (a.name.toLowerCase() < b.name.toLowerCase()) return -1;
+
+    if (a.name > b.name) return +1;
+    if (a.name < b.name) return -1;
+
+    return 0;
+  });
+
+  return sortedCs;
+}
+
+const joinClub = async (
+  name: string,
+  countMembers: number,
+  searchPreference?: boolean,
+): Promise<void> => {
+  await japi('post', '/join-club', { name });
+
+  const existingClubs = lastEvent<ClubItem[]>('updated-clubs') ?? [];
+
+  const updatedClubs = [
+    ...existingClubs
+      .filter((c) => c.name !== name)
+      .map((c) => ({
+        ...c,
+        search_preference:
+          searchPreference === true ? false : c.search_preference
+      })),
+    {
+      name,
+      count_members: countMembers,
+      search_preference: searchPreference
+    },
+  ];
+
+  sortClubs(updatedClubs)
+
+  notify<ClubItem[]>('updated-clubs', updatedClubs);
+};
+
+export {
+  ClubItem,
+  joinClub,
+  sortClubs,
+};
